Live Detection
Live detection working. Bluetooth and meshLoader are deactivated for
ease of testing. Vision needs to be narrowed, also needs testing with
belt. Should override front raycast if something is detected (although I
think I may have put that code in the wrong place. Have to go back to
raycasting script and check)

+++ New Tango/Assets/TangoSDK/Examples/PointToPoint/Scripts/PointToPointGUIController.cs
@@ -190,7 +190,7 @@ public class PointToPointGUIController : MonoBehaviour, ITangoDepth
             TangoEnums.TangoDepthCameraRate.DISABLED);
 
         Camera cam = Camera.main;
-        int pointIndex = m_pointCloud.FindClosestPoint(cam, touchPosition, 10);
+		int pointIndex = (int)m_pointCloud.FindClosestPoint(cam, touchPosition, 10);
 
         if (pointIndex > -1)
         {
